#cf Added CRUD for Workflows

+++ service_core/internal/usecase/usecase_workflow.go
@@ -12,13 +12,15 @@ import (
 type WorkflowlUsecase struct {
 	workflowRepo interfaces.WorkflowRepository
 	userRepo     interfaces.UserRepository
+	fileService  interfaces.FileService
 	log          *slog.Logger
 }
 
-func NewWorkflowUsecase(workflowRepo interfaces.WorkflowRepository, userRepo interfaces.UserRepository, log *slog.Logger) *WorkflowlUsecase {
+func NewWorkflowUsecase(workflowRepo interfaces.WorkflowRepository, userRepo interfaces.UserRepository, fileService interfaces.FileService, log *slog.Logger) *WorkflowlUsecase {
 	return &WorkflowlUsecase{
 		workflowRepo: workflowRepo,
 		userRepo:     userRepo,
+		fileService:  fileService,
 		log:          log,
 	}
 }
@@ -38,6 +40,7 @@ func (workflowlUsecase *WorkflowlUsecase) GetWorkflows(ctx context.Context, user
 	log.Debug("getting worklows from database")
 	workflows, err := workflowlUsecase.workflowRepo.GetWorkflows(ctx)
 	if err != nil {
+		// TODO: custom errors?
 		log.Error("failed to get workflows", slogger.Err(err))
 		return nil, fmt.Errorf("%s: %w", op, err)
 	}
@@ -48,6 +51,7 @@ func (workflowlUsecase *WorkflowlUsecase) GetWorkflows(ctx context.Context, user
 
 func (workflowlUsecase *WorkflowlUsecase) CreateWorkflow(ctx context.Context, name string, stages []domain.WorkflowStage, userID uint) error {
 	const op = "usecase.workflow.CreateWorkflow"
+
 	log := workflowlUsecase.log.With(slog.String("op", op))
 	log.Info("creating workfow")
 
@@ -58,28 +62,14 @@ func (workflowlUsecase *WorkflowlUsecase) CreateWorkflow(ctx context.Context, na
 	}
 
 	log.Debug("checking users from request")
-	userMap := make(map[uint]struct{})
-	for _, stage := range stages {
-		userMap[stage.UserID] = struct{}{}
-	}
-
-	userIDs := make([]uint, 0, len(userMap))
-	for userID := range userMap {
-		userIDs = append(userIDs, userID)
-	}
-
-	allExists, err := workflowlUsecase.userRepo.CheckUsersExist(ctx, userIDs)
-	if err != nil {
-		log.Error("failed users exist check", slogger.Err(err))
+	if err := workflowlUsecase.checkUsers(ctx, stages); err != nil {
+		log.Error("failed to validate users", slogger.Err(err))
 		return fmt.Errorf("%s: %w", op, err)
-	}
-	if !allExists {
-		log.Warn("some users are not exist", slogger.Err(domain.ErrUserNotFound))
-		return fmt.Errorf("%s: %w", op, domain.ErrUserNotFound)
 	}
 
 	log.Debug("putting workflow into db")
 	if err := workflowlUsecase.workflowRepo.CreateWorkflow(ctx, name, stages); err != nil {
+		// TODO: custom errors?
 		log.Error("failed to create workflow", slogger.Err(err))
 		return fmt.Errorf("%s: %w", op, err)
 	}
@@ -88,12 +78,44 @@ func (workflowlUsecase *WorkflowlUsecase) CreateWorkflow(ctx context.Context, na
 	return nil
 }
 
-func (workflowlUsecase *WorkflowlUsecase) UpdateWorkflow(ctx context.Context, workflowInfo []domain.Workflow, userID uint) error {
+func (workflowlUsecase *WorkflowlUsecase) UpdateWorkflow(ctx context.Context, workflowID uint, name string, stages []domain.WorkflowStage, userID uint) error {
+	const op = "usecase.workflow.UpdateWorkflow"
+
+	log := workflowlUsecase.log.With(slog.String("op", op))
+	log.Info("updating workfow")
+
+	log.Debug("checking if user is admin")
+	if err := workflowlUsecase.checkAdmin(ctx, userID); err != nil {
+		log.Error("failed admin check", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	log.Debug("checking workflow existence")
+	if err := workflowlUsecase.checkWorkflow(ctx, workflowID); err != nil {
+		log.Error("failed workflow existence check", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	log.Debug("checking users from request")
+	if err := workflowlUsecase.checkUsers(ctx, stages); err != nil {
+		log.Error("failed to validate users", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	log.Debug("updating workflow")
+	if err := workflowlUsecase.workflowRepo.UpdateWorkflow(ctx, workflowID, name, stages); err != nil {
+		// TODO: custom errors?
+		log.Error("failed to update workflow", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	log.Info("workflow updated successfully")
 	return nil
 }
 
 func (workflowlUsecase *WorkflowlUsecase) DeleteWorkflow(ctx context.Context, workflowID uint, userID uint) error {
 	const op = "usecase.workflow.DeleteWorkflow"
+
 	log := workflowlUsecase.log.With(slog.String("op", op))
 	log.Info("deleting workfow")
 
@@ -103,8 +125,27 @@ func (workflowlUsecase *WorkflowlUsecase) DeleteWorkflow(ctx context.Context, wo
 		return fmt.Errorf("%s: %w", op, err)
 	}
 
+	log.Debug("checking workflow existence")
+	if err := workflowlUsecase.checkWorkflow(ctx, workflowID); err != nil {
+		log.Error("failed workflow existence check", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	log.Debug("checking if this workflow in use")
+	exists, err := workflowlUsecase.fileService.CheckWorkflow(ctx, workflowID)
+	if err != nil {
+		// TODO: custom errors
+		log.Error("failed workflow check", slogger.Err(err))
+		return fmt.Errorf("%s: %w", op, err)
+	}
+	if exists {
+		log.Error("workflow is in use", slogger.Err(domain.ErrWorkflowInUse))
+		return fmt.Errorf("%s: %w", op, domain.ErrWorkflowInUse)
+	}
+
 	log.Debug("deleting workflow")
 	if err := workflowlUsecase.workflowRepo.DeleteWorkflow(ctx, workflowID); err != nil {
+		// TODO: custom errors
 		log.Error("failed to delete workflow", slogger.Err(err))
 		return fmt.Errorf("%s: %w", op, err)
 	}
@@ -119,6 +160,39 @@ func (workflowlUsecase *WorkflowlUsecase) checkAdmin(ctx context.Context, userID
 	}
 	if role != "admin" {
 		return domain.ErrAccessDenied
+	}
+	return nil
+}
+
+func (workflowlUsecase *WorkflowlUsecase) checkUsers(ctx context.Context, stages []domain.WorkflowStage) error {
+	userMap := make(map[uint]struct{})
+	for _, stage := range stages {
+		userMap[stage.UserID] = struct{}{}
+	}
+
+	userIDs := make([]uint, 0, len(userMap))
+	for userID := range userMap {
+		userIDs = append(userIDs, userID)
+	}
+
+	allExists, err := workflowlUsecase.userRepo.CheckUsersExist(ctx, userIDs)
+	if err != nil {
+		return err
+	}
+	if !allExists {
+		return domain.ErrUserNotFound
+	}
+
+	return nil
+}
+
+func (workflowlUsecase *WorkflowlUsecase) checkWorkflow(ctx context.Context, workflowID uint) error {
+	exists, err := workflowlUsecase.workflowRepo.CheckWorkflow(ctx, workflowID)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		return domain.ErrWorkflowNotFound
 	}
 	return nil
 }
